♻️  zb: Split basic tests into own mod
